JIRA-2987. Couple of problems - firstly we had some formatting of dates using mm/dd/yyyy which is not correct for 21st century software. It needs to be MM/dd/yyyy (lowercase mm returns the minute).
Second problem is hoofing arrays around arrays and swapping between types. Ideally we should have strongly typed arrays/collections but its doesn't seem to like passing arrays of type long/integer when they are multi-dimensions (or more annoyingly occasionally multi-dimensional)

Third problem was missing the .value off certain objects which meant the recordset was getting passed into functions rather than the actual value of the field. This wasn't a problem when the code was running under VB6 because it had default properties, but .NET requires these properties to be specified.

+++ Components/COAInt_Server/General.vb
@@ -40,7 +40,7 @@ Friend Class clsGeneral
 
     Dim i As Short
 
-    ' eg. DateFormat = "mm/dd/yyyy"
+    ' eg. DateFormat = "MM/dd/yyyy"
     '     Calendar   = "dd/mm/yyyy"
     '     DateString = "06/02/2000"
     '     Compare to = 02/06/2000
@@ -254,7 +254,7 @@ ErrorTrap:
     Dim rsTemp As ADODB.Recordset
     Dim strSQL As String
     Dim fOK As Boolean
-    Dim lngViews() As Integer
+    Dim lngViews() As Long
 
     ReDim lngViews(0)
